Create Virtual Stack Machine
Honestly made no sense for this to exist in the Skill Class.

We can pass in what we need to here and then work from there. What's
great about this approach is that we can pass our info directly to the
function rather than set it as an object property where it could get
messy.
